refactor(phase4): extract policy_parsing from eon_parser.rs
Move parse_row_policies and all filter/assignment/exists policy
helpers (623 lines) out of eon_parser.rs (9,559 → 8,940 lines)
into eon_parser/policy_parsing.rs (648 lines).

The two serde-attribute helpers (default_admin_bypass,
deserialize_create_policies_document) that documents.rs uses are
pub(super) in policy_parsing; documents.rs imports them via
`use super::policy_parsing::`.
